Looping Music test
Changes:
--created a new scene called 'morrowGroveMusicTestscene' that sloppily
uses the new Morrow Grove background from most recent update
--added an object called 'bgmPlayer'
--added a script to bgmPlayer called 'bgmMergeClipsAndLoop.cs' to the
Scripts folder that loads and loops music
--imported 2 (badly edited) clips from Morrow Grove BGM track in .wav
format into the bgm & sfx folder; they are used in bgmPlayer and its
script

As far as I know the new object and script shouldn't interfere with
anything else in the project.
Still many problems in need of fixing; check the comments at the top of
the new script for details and a TODO list.