BUUUUGGGG FIXES
shooting cooldown - lasers don't stay on screen anymore if you button
mash

added lock, unlock, and teleportable methods to the platformData script
for use in the future.  Tried implementing them, felt no need to -
commented out.

shifted platform spawns over a bit when the boss is on screen so that
there's less time that platforms spend inside the boss.